requirements.txt
discord.py
python-dotenv
openai
requests
beautifulsoup4
pytz
Pillow
geopy
termcolor
